Apply binary tree algorithm to create a maze from a grid
A maze can be created like so:

```
iex(1)> alias Twister.{Grid, BinaryTree}
iex(2)> grid = Grid.new(5, 4)
iex(24)> maze = BinaryTree.on(grid)
```

The binary tree has no concept of grid boundaries. It will link outside
the boundaries of the grid. This will be easier to fix when the maze can
be graphically viewed.

Note: this commit marks the switch from using Fossil for source control
to using Git. All previous commits were made in Fossil and were exported
with `fossil git export`.

The next task is to print the maze to stdout as ASCII art.