Do not link cells through grid boundaries
In cases where the cell being linked is at the eastern grid border the
binary tree will always link south or east when the cell is at the
southern border.

The exception is the south eastern corner of the grid which will not be
linked at all because there are no more cells south or east.