Partial solution for Sidewinder algorithm
This is as far as I got with the current implementation. It's quite
difficult to collect a random number of cells, link them, collect another
random run of cells from the remaining row, and keep track of all this
in my head.

The grid is represented as a struct with a map of cells. I am interested
in representing the cells with Erlang Term Storage instead to see if it
is easier to think about and use.